Fix import path changed because of refactoring
