--- pyopenapi/__init__.py
@@ -1,8 +1,0 @@
-from .getter import Getter
-from .core import App, Security
-
-
-# backward compatible
-SwaggerApp = App
-SwaggerSecurity = Security
-SwaggerAuth = SwaggerSecurity
