+++ pyopenapi/migration/versions/v1_2/scanner/__init__.py
@@ -1,0 +1,1 @@
+from .upgrade import Upgrade
