+++ pyopenapi/migration/versions/v3_0_0/__init__.py
@@ -1,0 +1,1 @@
+
