+++ pyopenapi/migration/versions/v3_0_0/objects.py
@@ -1,5 +1,5 @@
 from __future__ import absolute_import
-from ..base2 import Base2, field, rename, child, list_, map_
+from ...spec import Base2, field, rename, child, list_, map_
 import six
