+++ pyopenapi/migration/versions/v3_0_0/scanner/resolve.py
@@ -1,8 +1,8 @@
 from __future__ import absolute_import
-from ...utils import jp_compose, jr_split
-from ...errs import ReferenceError
-from ...scan import Dispatcher
-from ...spec.v3_0_0.objects import (
+from ....utils import jp_compose, jr_split
+from ....errs import ReferenceError
+from ....scan import Dispatcher
+from ..objects import (
     PathItem,
     Schema,
     Parameter,
