+++ pyopenapi/tests/migration/test_store.py
@@ -1,7 +1,7 @@
-from pyopenapi.store import SpecObjStore
-from pyopenapi.spec.v2_0.objects import Swagger, Info
-from pyopenapi.spec.v3_0_0.objects import OpenApi
-from .utils import get_test_file
+from pyopenapi.migration.store import SpecObjStore
+from pyopenapi.migration.versions.v2_0.objects import Swagger, Info
+from pyopenapi.migration.versions.v3_0_0.objects import OpenApi
+from ..utils import get_test_file
 import unittest
 import json
