+++ pyopenapi/tests/migration/versions/v3_0_0/test_load.py
@@ -1,6 +1,6 @@
-from pyopenapi import App
-from ..utils import get_test_file
-from ...spec.v3_0_0.objects import OpenApi
+from pyopenapi.contrib.pyswagger import App
+from pyopenapi.migration.versions.v3_0_0.objects import OpenApi
+from ....utils import get_test_file
 import unittest
 import os
 import yaml
